Replace BootableServiceProviderInterface with BootstrapperInterface
The reason for that change is that a bootstrapper may have dependencies.
These dependencies should be provided by container to the construct
method.  Since container is only be ready to use after registering all
service providers, it's not possible to implement the bootstrap and
register methods in a single class